drivers: dvfs: Fix the type and comment for dvfs_opp.rate
The existing SCPI function assume the rate is in MHz. This is a safe
assumption, and it allows shrinking the type (and thus the struct). So
change the type and document the correct expected unit.

+++ include/drivers/dvfs.h
@@ -16,7 +16,7 @@
 	(&container_of((dev)->drv, struct dvfs_driver, drv)->ops)
 
 struct dvfs_opp {
-	const uint32_t rate;    /**< Clock rate in Hz. */
+	const uint16_t rate;    /**< Clock rate in MHz. */
 	const uint16_t voltage; /**< Supply voltage in mV. */
 };
